Setup Functions

The main functionality provided by InputFiles is the ability to consistently setup packages from a .toml file.

Setup Functions

Usage

Global Options

Toml Options

Change default paths

Change logging behaviour

InputFiles Options

Add new global paths

API

Public objects

Private objects

Usage

Most of my packages use the same boilerplate to setup everything in a way that allows easy access propogation of user provided parameters:

using TOML
using InputFiles
using ArgParse

function get_args()
    s = ArgParseSettings()
    @add_arg_table s begin
        "--verbose", "-v"
            help = "Increase level of logging verbosity"
            action = :store_true
        "input"
            help = "Path to .toml file"
            required = true
    end
    return parse_args(s)
end

function main()
    args = get_args()
    verbose = args["verbose"]
    toml_path = args["input"]
    toml = TOML.parsefile(abspath(toml_path))
    setup_global!(toml, verbose, toml_path)
    @debug "Read in $toml_path"
    run_MyPackage(toml) # Actually run everything with the prepared `toml`
end

Given a .toml file, this will produce a dictionary with a global key containing a dictionary with

toml_path: The path to the .toml input file

base_path: A base path defined relative to toml_path from which any relative paths used by the package will be relative to

output_path: The output directory where all output, including logging, will be put

logging: Whether logging was set up. This allows you to use my InputFiles logging function or your own

log_file: If using my logging function, then in addition to stdout, logs will be saved to log_file

Global Options

You can change much of the behaviour of InputFiles setup logic both in how you call it, and in global options in the input .toml file.

Toml Options

Much of the behaviour can be controlled from the .toml file itself, without needed to change any code.

Change default paths

By default, InputFiles sets base_path equal to toml_path, and output_path equal to base_path/Output. By defining your own (absolute or relative) base_path, and output_path, you can change these defaults. Note that base_path is defined relative to toml_path, and output_path is defined relative to base_path, but absolute paths are respected.

[ global ]
    base_path = "../Examples"
    # base_path = toml_path/../Examples
    output_path = "/tmp/test"
    # output_path = "/tmp/test"

Change logging behaviour

You can define in the .toml file, whether you want to use InputFiles logging or not. Additionally you can specify a new location for the log file (defaults to output_path/log.txt). This can be either absolute, or relative to output_path

[ global ]
    logging = false
    log_file = logs/out.txt
    # log_file = output_path/logs/out.txt

InputFiles Options

Add new global paths

By defining a Dict of paths, you can add new global paths, which can be accessed throughout the package. The following creates data_path relative to base_path, and filter_path relative to data_path. You must provide defaults for these paths but the user can overwrite then in the .toml (see Change default paths).

paths = Dict(
    "data_path" => ("base_path", "Data")
    "filter_path" => ("data_path", "Filters")

setup_global!(toml, toml_path, paths)

See default_paths for more details.

API

Public objects

These are the functions, types, and other objects you have access to with a simple using InputFiles

InputFiles.SetupModule.setup_global! — Function

setup_global!(input::Dict, input_path::AbstractString, verbose::Bool, paths::OrderedDict{String, Tuple{String, String}}=OrderedDict{String, Tuple{String, String}}(), log_path::String="output_path"; test::Bool=false)

Setup the "global" information of input, including paths and logging.

Arguments

input::Dict: The input file loaded into a Dict

input_path::AbstractString: The path to the input .input file (from which input is loaded)

verbose::Bool: Whether or not to display @debug calls

paths::OrderedDict{String, Tuple{String, String}}: Paths to expand. paths will be merged with the following default_paths, with paths taking preference. See default_paths for a the syntax of paths

output_path::String: The "path_name" of the directory where logging should output

test::Bool: Whether this run is a test or not. This will decide whether to actually create directories, or log anything

See also setup_paths!, and setup_logging!

Example

input = Dict()
input_path = "/path/to/input/file.input"
setup_global!(input, input_path, false; test=true)
println(input)

# output

Dict{Any, Any}("global" => Dict{Any, Any}("input_path" => "/path/to/input", "logging" => false, "base_path" => "/path/to/input/", "log_file" => "/path/to/input/Output/log.txt", "output_path" => "/path/to/input/Output"))

Private objects

These are the functions, types, and other objects which you don't have access to by default. They are included here for verbosity and developer assistance.

InputFiles.SetupModule.default_paths — Constant

default_paths::OrderedDict{String, Tuple{String, String}}

The default paths which will be expanded into absolute paths and used throughout the project.

default_paths::OrderedDict{String, Tuple{String, String}} = OrderedDict{String, Tuple{String, String}}(
    # Name => relative, default
    "base_path" => ("input_path", ""),
    "output_path" => ("base_path", "Output")
)

This dictionary maps ("path_name" => ("relative_name", "default_path")), where "path_name" is a human readable name for the path, "relative_name" is the name of the path which "path_name" is relative to, and "default_path" is the default value for the path (either absolute or relative). If "path_name" already exists inside input["global"], then that path will be used either as is (if an absolute path) or relative to "relative_name", otherwise the "default_path" will be used

InputFiles.SetupModule.setup_logger — Method

setup_logger(log_file::AbstractString, verbose::Bool)

Helper function which sets up Logger formating and log level

Arguments

log_file::AbstractString: Path to log file

verbose::Bool: Whether or not to display @debug calls

See also setup_logging! and setup_global!

InputFiles.SetupModule.setup_logging! — Function

setup_logging!(input::Dict, output_path::String="output_path"; log::Bool=true)

Helper function which sets up log level, log files, etc... Assumes that setup_paths! has already been run on input

Arguments

input::Dict: The input file we are modifying. This assumes that the input already has a "global" key with value Dict{Any, Any}("input_path"=>"/path/to/input"), where"input_path"` is the path to the input file

output_path::String: The "path_name" of the output directory where log files should be written. See default_paths for more details

See also setup_global!

InputFiles.SetupModule.setup_paths! — Method

setup_paths!(input::Dict, paths::OrderedDict{String, Tuple{String, String}})

Helper function which sets up paths, expanding relative paths and ensuring all interim directories exist

Arguments

input::Dict: The input file we are modifying. This assumes that the input already has a "global" key with value Dict{Any, Any}("input_path"=>"/path/to/input"), where"input_path"is the path to theinput` file

paths::OrderedDict{String, Tuple{String, String}}: Paths to expand. paths will be merged with the following default_paths, with paths taking preference. See default_paths for a the syntax of paths

See also setup_global!